feat: Add Terraform configuration for creating a GKE cluster
This commit adds the necessary Terraform configuration files to create a Google Kubernetes Engine (GKE) cluster. The main.tf file includes the provider configuration for Google Cloud and a random provider for generating a suffix for the cluster name. It also defines a random_string resource to generate a random suffix for the cluster name. The google_container_cluster resource is used to create the primary cluster with autoscaling enabled. The node_pool resource defines a preemptible node pool with autoscaling. The outputs.tf file includes the output variables for the cluster name, cluster endpoint, and master version.

The variables.tf file contains the input variables for the project ID, region, cluster name, machine type, minimum number of nodes, and maximum number of nodes.

Signed-off-by: Ihor Dvoretskyi <ihor@linux.com>